+++ src/components/top/index.jsx
@@ -1,8 +1,13 @@
 import React from 'react'
 import { Link } from 'gatsby'
 import { GitHubIcon } from '../social-share/github-icon'
+import Search from '../search'
 
 import './index.scss'
+
+const searchIndices = [
+  { name: `Posts`, title: `Blog Posts`, hitComp: `PostHit` },
+]
 
 export const Top = ({ title, location, rootPath }) => {
   const isRoot = location.pathname === rootPath
@@ -13,6 +18,7 @@ export const Top = ({ title, location, rootPath }) => {
           {title}
         </Link>
       )}
+      <Search collapse indices={searchIndices} />
       <GitHubIcon />
     </div>
   )
